+++ cli/client/generic.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"reflect"
 
+	"github.com/go-resty/resty/v2"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/types"
 )
@@ -131,6 +132,21 @@ func (client *RestClient) Post(path string, obj interface{}) error {
 	}
 
 	return nil
+}
+
+// PostWithResponse sends a POST request with obj as the payload to the given path
+// and additionally returns the response
+func (client *RestClient) PostWithResponse(path string, obj interface{}) (*resty.Response, error) {
+	res, err := client.R().SetBody(obj).Post(path)
+	if err != nil {
+		return nil, err
+	}
+
+	if res.StatusCode() >= 400 {
+		return nil, UnmarshalError(res)
+	}
+
+	return res, nil
 }
 
 // Put sends a PUT request with obj as the payload to the given path